requirements.txt
Django==4.2.29
djangorestframework==3.15.1
drf-nested-routers==0.93.5
django-cors-headers==4.3.1
python-decouple==3.8
psycopg2-binary==2.9.9
